# Raw JSON writer (~10x faster) (#5314) 