… arrow-json/test/data/basic.json