# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #5314

# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

```
bench_primitive         time:   [2.9493 ms 2.9504 ms 2.9515 ms]
                        change: [-86.148% -86.054% -85.962%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

bench_mixed             time:   [6.0997 ms 6.1016 ms 6.1038 ms]
                        change: [-86.966% -86.905% -86.856%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  6 (6.00%) high mild
  1 (1.00%) high severe

bench_struct            time:   [7.4146 ms 7.4169 ms 7.4193 ms]
                        change: [-89.586% -89.536% -89.485%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild

bench_nullable_struct   time:   [2.3035 ms 2.3052 ms 2.3070 ms]
                        change: [-91.153% -91.131% -91.109%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

bench_list              time:   [2.4907 ms 2.4960 ms 2.5011 ms]
                        change: [-91.988% -91.965% -91.942%] (p = 0.00 < 0.05)
                        Performance has improved.

bench_nullable_list     time:   [982.97 µs 983.21 µs 983.46 µs]
                        change: [-85.818% -85.807% -85.795%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  4 (4.00%) high severe

Benchmarking bench_struct_list: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.9s, enable flat sampling, or reduce sample count to 40.
bench_struct_list       time:   [1.9682 ms 1.9691 ms 1.9699 ms]
                        change: [-89.756% -89.726% -89.700%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->