@tustvold 2024-01-23
I will re-run the benchmarks tomorrow